move the last couple of things
