+++ src/holofuel/components/layout/PrimaryLayout/PrimaryLayout.js
@@ -3,8 +3,8 @@ import { object } from 'prop-types'
 import cx from 'classnames'
 import ScreenWidthContext from 'contexts/screenWidth'
 import styles from './PrimaryLayout.module.css' // eslint-disable-line no-unused-vars
-import 'global-styles/colors.css'
-import 'global-styles/index.css'
+import 'holofuel/global-styles/colors.css'
+import 'holofuel/global-styles/index.css'
 
 import Header from 'holofuel/components/Header'
