+++ src/index.js
@@ -5,7 +5,7 @@ import { BrowserRouter as Router } from 'react-router-dom'
 import { useMediaPredicate } from 'react-media-hook'
 import apolloClient from 'apolloClient'
 import ReactModal from 'react-modal'
-import HFRouter from './HFRouter'
+import HFRouter from './holofuel/HFRouter'
 import RegisterUser from 'components/RegisterUser'
 import ScreenWidthContext from 'contexts/screenWidth'
 import HPAdminRouter from './HPAdminRouter'
